requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy==1.22.4
pandas
scikit-learn==1.1.1
pandas-ta
matplotlib
seaborn
tensorflow==2.8.0

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs
yfinance
python-binance

# utilities
six>=1.14
joblib
memoized-property
termcolor
python-dotenv
